requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.14.0
yfinance==0.2.22
alpha-vantage==2.3.1
tweepy==4.14.0
textblob==0.17.1
vaderSentiment==3.3.2
python-dotenv==1.0.0
ccxt==4.1.22
websocket-client==1.6.4
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==2.3.7
plotly==5.17.0
ta==0.10.2
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
python-binance==1.0.19
newsapi-python==0.2.7